Framework resolution tests (dotnet/core-setup#5502)
* Framework resolution tests

Backfill detailed tests for framework resolution behavior, specifically
around roll-forward.

Introduces some new test helpers
* hostpolicy mock, to avoid actually loading and running coreclr just to
validate selected frameworks
* Builder-like API for `.runtimeconfig.js`
* Builder-like API for shared frameworks
* Some helper classes to limit file copy while keeping the tests fully
isolated


Commit migrated from https://github.com/dotnet/core-setup/commit/e42109d82608c0c0a6123f9b70c21ab632b08d07

+++ src/installer/test/TestUtils/TestArtifact.cs
@@ -47,7 +47,7 @@ namespace Microsoft.DotNet.CoreSetup.Test
 
         public void Dispose()
         {
-            if (!PreserveTestRuns())
+            if (!PreserveTestRuns() && Directory.Exists(Location))
             {
                 Directory.Delete(Location, true);
             }
